refactor(content): share the select choice helper
